requirements.txt
a2wsgi==1.10.10
aiosmtplib==5.1.0
aiosqlite==0.21.0
alembic==1.18.3
annotated-types==0.7.0
anyio==4.12.1
argcomplete==3.6.3
asgiref==3.11.1
attrs==25.4.0
babel==2.18.0
cadwyn==5.4.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.7
colorlog==6.10.1
configparser==7.2.0
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.4
Deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.20
fsspec>=2026.2.0
googleapis-common-protos==1.72.0
greenback==1.3.0
greenlet==3.3.1
grpcio==1.78.0
great-expectations==0.18.21
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgspec==0.20.0
natsort==8.4.0
numpy==1.26.4
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
outcome==1.3.0.post0
packaging==26.0
pandas==2.2.3
pathspec==1.0.4
pendulum==2.1.2
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
protobuf==6.33.5
psutil==7.2.2
pyarrow==10.0.1
pyspark==3.5.5
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.11.0
minio==7.2.18
psycopg2-binary==2.9.11
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.3.2
rich-argparse==1.7.2
rich-toolkit==0.18.1
rpds-py==0.30.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
starlette==0.48.0
structlog==25.5.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.4
termcolor==3.3.0
text-unidecode==1.3
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
universal_pathlib==0.2.6
update-checker==0.18.0
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wirerope==1.0.0
wrapt==2.1.1
zipp==3.23.0
holidays==0.58
# Machine Learning and Feature Engineering
xgboost==2.0.3
scikit-learn==1.4.1.post1
s3fs>=2026.2.0
